feat
1. Refactorización del Carrito (Carrito/)
Antes era un archivo muy denso, ahora es una estructura modular:

Productos.js / .css: Se creó este componente para manejar exclusivamente la lista de productos y la selección de "todo".
Resumen.js / .css: Se creó este componente para manejar el panel lateral con los totales y el botón de compra.
Carrito.js: Se simplificó para actuar solo como el contenedor principal que une las dos piezas.
2. Refactorización Profesional de TarjetaPago (TarjetaPago/)
Este fue el cambio más profundo. Pasamos de un archivo de 500 líneas a una micro-arquitectura:

TarjetaPago.js: Ahora es un componente visual muy ligero (menos de 100 líneas).
useTarjetaPago.js (Custom Hook): Extrajimos todo el "cerebro" (estados, validaciones, proceso de pago) a este archivo.
utilsTarjeta.js: Movimos la lógica matemática de validación de tarjetas y formateo aquí.
Sub-componentes Modulares:
VistaPreviaTarjeta: Solo la parte visual de la tarjeta azul/oscura.
FormularioTarjeta: Los inputs de nombre, número y CVV.
SeccionCupon: La lógica de cupones (ahora con estilos CSS reales, sin estilos en línea).
ModalExitoPago: El mensaje de confirmación final.
