--- client/marketplace-frontend/src/componentes/PasarelaDePago/TarjetaPago.css
@@ -1,324 +1,0 @@
-/* ============================================
-   Card Payment Styles - CardPay Component
-   ============================================ */
-
-.pasarela-de-pago {
-    background: var(--nexora-card, rgba(14, 25, 45, 0.7));
-    backdrop-filter: blur(15px);
-    border-radius: 20px;
-    padding: 30px;
-    border: 1px solid var(--nexora-border, rgba(255, 255, 255, 0.05));
-    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.3);
-    max-width: 600px;
-    margin: 0 auto;
-    width: 100%;
-}
-
-.titulo-pago-pasarela {
-    font-size: 1.4rem;
-    color: white;
-    margin-bottom: 8px;
-    display: flex;
-    align-items: center;
-    gap: 12px;
-}
-
-.subtitulo-pago-pasarela {
-    color: #9ca3af;
-    font-size: 0.9rem;
-    margin-bottom: 25px;
-}
-
-/* Layout de dos columnas dentro de pasarela */
-.contenido-tarjeta-dos-columnas {
-    display: flex;
-    gap: 25px;
-    margin-bottom: 25px;
-}
-
-.columna-izquierda-tarjeta {
-    flex: 1;
-    min-width: 250px;
-}
-
-.columna-derecha-formulario {
-    flex: 1;
-    min-width: 250px;
-}
-
-/* Card Preview */
-.vista-previa-tarjeta-pasarela {
-    background: linear-gradient(135deg, #1e3a5f 0%, #0f172a 100%);
-    border-radius: 12px;
-    padding: 20px;
-    min-height: 130px;
-    position: relative;
-    overflow: hidden;
-    transition: all 0.4s ease;
-    border: 1px solid rgba(255, 255, 255, 0.1);
-    margin-right: -20px;
-    margin-left: -20px;
-
-}
-
-.chip-tarjeta-pasarela {
-    width: 45px;
-    height: 35px;
-    background: linear-gradient(135deg, #fbbf24, #d97706);
-    border-radius: 5px;
-    flex-shrink: 0;
-    margin-bottom: 15px;
-}
-
-.vista-previa-numero-tarjeta {
-    font-family: 'Courier New', monospace;
-    font-size: 1.1rem;
-    color: white;
-    letter-spacing: 2px;
-    margin-bottom: 15px;
-    text-shadow: 0 2px 4px rgba(0,0,0,0.3);
-}
-
-.vista-previa-detalles-tarjeta {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-end;
-    gap: 10px;
-}
-
-.vista-previa-nombre-tarjeta {
-    font-size: 0.75rem;
-    color: rgba(255, 255, 255, 0.85);
-    text-transform: uppercase;
-    letter-spacing: 1px;
-}
-
-.vista-previa-vencimiento-tarjeta {
-    font-size: 0.75rem;
-    color: rgba(255, 255, 255, 0.7);
-}
-
-.vista-previa-logo-tarjeta {
-    position: absolute;
-    top: 15px;
-    right: 15px;
-    width: 50px;
-    height: 32px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-/* Form Styles */
-.formulario-pago-pasarela {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    margin-bottom: 0;
-}
-
-.entrada-pago-pasarela {
-    width: 100%;
-    padding: 12px 14px;
-    background: rgba(255, 255, 255, 0.05);
-    border: 1px solid var(--nexora-border, rgba(255, 255, 255, 0.1));
-    border-radius: 8px;
-    color: white;
-    font-size: 0.95rem;
-    transition: all 0.3s;
-}
-
-.fila-detalles-tarjeta-pasarela {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 12px;
-}
-
-.entrada-pago-pasarela:focus {
-    outline: none;
-    border-color: var(--nexora-blue, #0094ff);
-    background: rgba(255, 255, 255, 0.1);
-}
-
-.entrada-pago-pasarela.error-entrada {
-    border-color: #ef4444;
-}
-
-.grupo-entrada-pasarela {
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-}
-
-.campo-cvv-pasarela {
-    position: relative;
-}
-
-.alternar-cvv-pasarela {
-    position: absolute;
-    right: 12px;
-    top: 50%;
-    transform: translateY(-50%);
-    background: none;
-    border: none;
-    color: #9ca3af;
-    cursor: pointer;
-    padding: 5px;
-}
-
-.mensaje-error-pasarela {
-    color: #ef4444;
-    font-size: 0.75rem;
-    margin-top: 2px;
-}
-
-.error-general-pago {
-    background: rgba(239, 68, 68, 0.1);
-    color: #ef4444;
-    padding: 12px;
-    border-radius: 8px;
-    margin-bottom: 20px;
-    font-size: 0.9rem;
-    text-align: center;
-    border: 1px solid rgba(239, 68, 68, 0.2);
-}
-
-/* Buttons */
-.boton-enviar-pasarela {
-    width: 100%;
-    padding: 16px;
-    background: linear-gradient(135deg, var(--nexora-blue, #0094ff), var(--nexora-cyan, #00d4ff));
-    color: white;
-    border: none;
-    border-radius: 12px;
-    font-size: 1.1rem;
-    font-weight: 700;
-    cursor: pointer;
-    transition: all 0.3s;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.boton-enviar-pasarela:hover:not(:disabled) {
-    transform: translateY(-2px);
-    filter: brightness(1.1);
-    box-shadow: 0 4px 15px rgba(0, 148, 255, 0.3);
-}
-
-.boton-enviar-pasarela:disabled {
-    opacity: 0.6;
-    cursor: not-allowed;
-}
-
-.animacion-giro {
-    animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-    from { transform: rotate(0deg); }
-    to { transform: rotate(360deg); }
-}
-
-.advertencia-direccion {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    color: #f59e0b;
-    font-size: 0.85rem;
-    margin-top: 15px;
-    justify-content: center;
-}
-
-/* Success Modal */
-.capa-modal-exito-pasarela {
-    position: fixed;
-    top: 0; left: 0; right: 0; bottom: 0;
-    background: rgba(0,0,0,0.8);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 2000;
-    backdrop-filter: blur(5px);
-}
-
-.modal-exito-pasarela {
-    background: var(--nexora-card, #111827);
-    border-radius: 24px;
-    padding: 40px;
-    max-width: 500px;
-    width: 90%;
-    text-align: center;
-    border: 1px solid var(--nexora-border, rgba(255,255,255,0.1));
-    box-shadow: 0 25px 50px -12px rgba(0,0,0,0.5);
-}
-
-.icono-modal-exito-pasarela {
-    margin-bottom: 20px;
-    display: flex;
-    justify-content: center;
-}
-
-.detalles-modal-exito-pasarela {
-    margin: 25px 0;
-    text-align: left;
-}
-
-.info-tarjeta-modal-pasarela {
-    background: rgba(255,255,255,0.03);
-    padding: 15px;
-    border-radius: 12px;
-    margin-top: 15px;
-}
-
-.fila-info-modal-pasarela {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 10px;
-}
-
-.etiqueta-modal-pasarela { color: #9ca3af; font-size: 0.9rem; }
-.valor-modal-pasarela { color: white; font-weight: 600; }
-.monto-modal-pasarela { color: #10b981; font-size: 1.2rem; }
-
-.boton-modal-exito-pasarela {
-    width: 100%;
-    padding: 14px;
-    background: var(--nexora-blue, #0094ff);
-    color: white;
-    border: none;
-    border-radius: 12px;
-    font-weight: 700;
-    cursor: pointer;
-}
-
-/* Trust Section */
-.seccion-confianza-pasarela {
-    margin-top: 30px;
-    text-align: center;
-    border-top: 1px solid rgba(255,255,255,0.05);
-    padding-top: 20px;
-}
-
-.texto-pago-seguro-pasarela {
-    color: #10b981;
-    font-size: 0.85rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 8px;
-    margin-bottom: 15px;
-}
-
-.metodos-pago-pasarela {
-    display: flex;
-    justify-content: center;
-    gap: 15px;
-}
-
-.metodos-pago-pasarela span {
-    color: #4b5563;
-    font-size: 0.75rem;
-    text-transform: uppercase;
-    letter-spacing: 1px;
-}
