+++ client/marketplace-frontend/src/componentes/PasarelaDePago/TarjetaPago/SeccionCupon.css
@@ -1,0 +1,111 @@
+/* Coupon Section Styles */
+.seccion-cupon-pasarela {
+    margin: 16px 0;
+    padding: 16px;
+    background: rgba(168, 85, 247, 0.08);
+    border-radius: 14px;
+    border: 1px solid rgba(168, 85, 247, 0.2);
+}
+
+.etiqueta-cupon {
+    display: block;
+    font-size: 13px;
+    color: rgba(255, 255, 255, 0.7);
+    margin-bottom: 8px;
+    font-weight: 600;
+}
+
+.grupo-entrada-cupon {
+    display: flex;
+    gap: 10px;
+    margin-top: 8px;
+}
+
+.entrada-cupon {
+    flex: 1;
+    padding: 10px 14px;
+    background: rgba(255, 255, 255, 0.06);
+    border: 1px solid rgba(255, 255, 255, 0.12);
+    border-radius: 10px;
+    color: white;
+    font-size: 14px;
+    font-family: 'Inter', sans-serif;
+    letter-spacing: 1px;
+    transition: all 0.3s ease;
+}
+
+.entrada-cupon:focus {
+    outline: none;
+    border-color: rgba(168, 85, 247, 0.5);
+    background: rgba(255, 255, 255, 0.1);
+}
+
+.boton-validar-cupon {
+    padding: 10px 18px;
+    background: linear-gradient(135deg, #a855f7, #7c3aed);
+    border: none;
+    border-radius: 10px;
+    color: white;
+    cursor: pointer;
+    font-weight: 600;
+    font-size: 13px;
+    white-space: nowrap;
+    transition: all 0.3s ease;
+}
+
+.boton-validar-cupon:hover:not(:disabled) {
+    transform: translateY(-2px);
+    box-shadow: 0 4px 12px rgba(124, 58, 237, 0.3);
+    filter: brightness(1.1);
+}
+
+.boton-validar-cupon:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+}
+
+.boton-quitar-cupon {
+    padding: 10px 14px;
+    background: rgba(239, 68, 68, 0.15);
+    border: 1px solid rgba(239, 68, 68, 0.3);
+    border-radius: 10px;
+    color: #f87171;
+    cursor: pointer;
+    font-size: 13px;
+    font-weight: 500;
+    transition: all 0.3s ease;
+}
+
+.boton-quitar-cupon:hover {
+    background: rgba(239, 68, 68, 0.25);
+}
+
+.mensaje-error-cupon {
+    color: #f87171;
+    font-size: 12px;
+    margin-top: 8px;
+    padding-left: 4px;
+}
+
+.cupon-aplicado-info {
+    margin-top: 12px;
+    padding: 10px 14px;
+    background: rgba(16, 185, 129, 0.12);
+    border: 1px solid rgba(16, 185, 129, 0.3);
+    border-radius: 10px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.mensaje-exito-cupon {
+    color: #34d399;
+    font-weight: 600;
+    font-size: 14px;
+}
+
+.monto-descuento-cupon {
+    color: #34d399;
+    font-weight: 700;
+    font-size: 16px;
+}
